# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Tarefas",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/tasks": {
            "get": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Lista todas as tarefas",
                "description": "Retorna a lista de tarefas cadastradas no sistema.",
                "operationId": "f7353e9a7bb0ba4fcf0f2e8ba5ae41e8",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Define o formato de resposta esperado. Deve ser application/json",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "task_number",
                        "in": "query",
                        "description": "Filtrar pela número da tarefa",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filtrar por título ou descrição",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "responsible",
                        "in": "query",
                        "description": "Filtrar por responsável",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Em andamento",
                                "Concluída"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de tarefas recuperada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de tarefas recuperada com sucesso"
                                        },
                                        "data": {
                                            "properties": {
                                                "tasks": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "responsibles": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Cria uma nova tarefa",
                "description": "Cria uma nova tarefa no sistema.",
                "operationId": "acd076138bcbaeedf30ee2a12c656863",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Define o formato de resposta esperado. Deve ser application/json",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "responsible",
                                    "priority"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Nova Tarefa"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição da tarefa"
                                    },
                                    "responsible": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "alta",
                                            "média",
                                            "baixa"
                                        ],
                                        "example": "alta"
                                    },
                                    "deadline": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarefa criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tarefa criada com sucesso"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/tasks/{id}/complete": {
            "post": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Marca uma tarefa como concluída",
                "description": "Atualiza o status da tarefa para 'Concluída'.",
                "operationId": "3f009338be3af50211055fe2262463df",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Define o formato de resposta esperado. Deve ser application/json",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tarefa a ser marcada como concluída",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarefa concluída com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tarefa concluída com sucesso"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tasks/{id}/chatgpt": {
            "post": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Gera sugestão do ChatGPT para uma tarefa",
                "description": "Utiliza o título e a descrição da tarefa para gerar uma sugestão com o melhor jeito e o passo a passo prático para realizar a tarefa.",
                "operationId": "5ff38f84e950f8f86d8710e15ab22d86",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Define o formato de resposta esperado. Deve ser application/json",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tarefa para a qual a sugestão do ChatGPT será gerada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sugestão gerada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sugestão gerada com sucesso"
                                        },
                                        "data": {
                                            "properties": {
                                                "response": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarefa não encontrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}